pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "morag"
version = "0.1.0"
description = "Multimodal RAG Ingestion Pipeline"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi==0.115.4",
    "uvicorn==0.32.1",
    "python-multipart==0.0.17",
    "pydantic==2.10.3",
    "pydantic-settings==2.7.0",
    "pydantic-ai>=0.0.14",
    "celery==5.3.6",

    "redis==5.2.1",
    "qdrant-client==1.12.1",
    "google-genai==1.18.0",
    "yt-dlp==2024.12.13",
    "beautifulsoup4==4.12.3",
    "markdownify==0.14.1",


    "markdown==3.7",

    "aiofiles==24.1.0",
    "httpx==0.28.1",
    "python-dotenv==1.0.1",
    "structlog==24.4.0",
    "psutil==6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.0.0,<7.0.0",
    "mypy>=1.7.0,<2.0.0",
    "pre-commit>=3.5.0,<4.0.0",

]

docling = [
    "docling>=2.36.0",

]

nlp = [
    "spacy>=3.7.0,<4.0.0",
    "unstructured>=0.11.0,<1.0.0",
    "cloudpathlib>=0.20.0,<1.0.0",

]

audio = [
    "faster-whisper>=1.0.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "mutagen>=1.47.0",
    "pyannote.audio>=3.1.0",
    "sentence-transformers>=2.2.0",
]

video = [
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "torch>=2.0.0",
]

image = [
    "Pillow>=10.0.0,<11.0.0",
    "pytesseract>=0.3.10,<1.0.0",
    "easyocr>=1.7.0,<2.0.0",
    "opencv-python>=4.8.0,<5.0.0",
]

office = [
    "python-docx>=1.1.2,<2.0.0",
    "openpyxl>=3.1.5,<4.0.0",
    "python-pptx>=0.6.21,<1.0.0",
]

web = [
    "playwright>=1.40.0,<2.0.0",
    "trafilatura>=1.6.0,<2.0.0",
]

all-extras = [
    "docling>=2.7.0,<3.0.0",
    "spacy>=3.7.0,<4.0.0",
    "unstructured>=0.11.0,<1.0.0",
    "cloudpathlib>=0.20.0,<1.0.0",
    "faster-whisper>=1.0.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "mutagen>=1.47.0",
    "pyannote.audio>=3.1.0",
    "sentence-transformers>=2.2.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0,<11.0.0",
    "torch>=2.0.0",
    "pytesseract>=0.3.10,<1.0.0",
    "easyocr>=1.7.0,<2.0.0",
    "python-docx>=1.1.2,<2.0.0",
    "openpyxl>=3.1.5,<4.0.0",
    "python-pptx>=0.6.21,<1.0.0",
    "playwright>=1.40.0,<2.0.0",
    "trafilatura>=1.6.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/morag"
Repository = "https://github.com/yourusername/morag"
Documentation = "https://github.com/yourusername/morag/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"